refactor(state): split openclaw-state-db into focused modules under the line limit
