+++ scripts/check-kysely-guardrails.mjs
@@ -41,8 +41,14 @@ const rawSqliteAllowPathGroups = {
     "src/state/openclaw-agent-db-session-migrations.ts",
     "src/state/openclaw-agent-db-session-provenance.ts",
     "src/state/openclaw-agent-db.ts",
+    "src/state/openclaw-state-db-audit-migration.ts",
+    "src/state/openclaw-state-db-legacy-backfills.ts",
+    "src/state/openclaw-state-db-maintenance.ts",
     "src/state/openclaw-state-db-operator-approval-migration.ts",
+    "src/state/openclaw-state-db-schema-additive.ts",
     "src/state/openclaw-state-db-schema-helpers.ts",
+    "src/state/openclaw-state-db-schema-repair.ts",
+    "src/state/openclaw-state-db-startup-checkpoint.ts",
     "src/state/openclaw-state-db.ts",
     "src/state/sqlite-schema-shape.test-support.ts",
   ],
